visp_lilo : Fix lilo mbox crash.
Mbox crash observed while receiving first mbox response in lilo drivers.
This crash is observed after 32bit variables are converted to
64-bit in visp/include/media_device.h

This issue is caused due to misaligned isp_dev structures in mbox.

Issue is resolved after adding same change in
visp_lilo/include/media_device.h file.

+++ visp_lilo/include/media_device.h
@@ -240,7 +240,7 @@ typedef struct
 
 typedef struct
 {
-	uint32_t DmaAddr;
+	uint64_t DmaAddr;
 	uint32_t DmaSize;
 } MediaPlane;
